+++ packages/core/package.json
@@ -17,11 +17,14 @@
     "src"
   ],
   "scripts": {
-    "build": "node ../../tools/scripts/build.mjs src/index.ts",
-    "dev": "node ../../tools/scripts/dev.mjs src/index.ts",
+    "build": "builder src/index.ts",
+    "dev": "builder src/index.ts --watch",
     "types": "tsc --emitDeclarationOnly --outDir dist --watch"
   },
   "dependencies": {
     "@threeify/math": "*"
+  },
+  "devDependencies": {
+    "builder": "*"
   }
 }
