+++ packages/scene/package.json
@@ -7,17 +7,18 @@
   "module": "./dist/index.js",
   "source": "./src/index.ts",
   "scripts": {
-    "build": "node ../../tools/scripts/build.mjs src/index.ts",
-    "dev": "node ../../tools/scripts/dev.mjs src/index.ts",
+    "build": "builder src/index.ts",
+    "dev": "builder src/index.ts --watch",
     "types": "tsc --emitDeclarationOnly --outDir dist --watch"
-  },
-  "devDependencies": {
-    "@types/draco3dgltf": "^1.4.0"
   },
   "dependencies": {
     "@threeify/core": "*",
     "@threeify/math": "*",
     "draco3dgltf": "^1.5.5",
     "meshoptimizer": "^0.18.1"
+  },
+  "devDependencies": {
+    "@types/draco3dgltf": "^1.4.0",
+    "builder": "*"
   }
 }
