+++ tools/builder/bin/cli.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { main } from '../dist/index.js';
+
+main();
